+++ ethereum/core/src/main/java/tech/pegasys/teku/core/signatures/SlashingProtectedSigner.java
@@ -13,6 +13,8 @@
 
 package tech.pegasys.teku.core.signatures;
 
+import java.net.URL;
+import java.util.Optional;
 import java.util.function.Consumer;
 import java.util.function.Supplier;
 import org.apache.tuweni.bytes.Bytes32;
@@ -140,8 +142,8 @@ public class SlashingProtectedSigner implements Signer {
   }
 
   @Override
-  public boolean isLocal() {
-    return delegate.isLocal();
+  public Optional<URL> getSigningServiceUrl() {
+    return delegate.getSigningServiceUrl();
   }
 
   @Override